backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
xarray==2023.11.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
scikit-learn==1.3.2
statsmodels==0.14.0
numpy==1.24.3
netcdf4==1.6.5
h5netcdf==1.2.0
earthaccess==0.11.0
aiohttp==3.9.1
httpx==0.25.2
aiofiles==23.2.1